This actually kicked my ass, I continue to fight on, however.
Restarted the entire thing, going to do some other stuff with it
like the distance visualized!

Hoping to add actual elastic collision detection (thats what killed me before, ijust couldnt get it working)

Main difference this time is im using Point which, to my understanding, is simply just a way to store and manipulate 2D coords. Specifically, im using Point2D which supports floating point integers. Im using it to represent the velocity of the balls.

And lastly, i thought that showing the triangle that forms when using the distance formula would be cool. It kinda is.